Preserve bitmap weight strength range checks
First divergence: an extreme public FT_GlyphSlot_AdjustWeight xdelta left C's bitmap slot unchanged, while Rust narrowed the 64-bit strength to i32, wrapped it to one pixel, and changed advance.x from 192 to 256.

FreeType keeps FT_Pos strengths through the FT_INT range checks in ftsynth.c and ftbitmap.c. Keep bitmap strengths as i64 until those checks, then pass validated pixel counts to the safe bitmap mutator. Add exact upper-x, upper-y, and lower-y rejection observations.

Impact: runtime parity remains 7079/7079 with 4 pending; real-parity routes remain 3874. api.rs coverage is 1041/1091 lines and 240/296 branches.

+++ pillow-rs-freetype/src/ffi/handles.rs
@@ -1577,10 +1577,11 @@ pub fn FT_GlyphSlot_AdjustWeight(
     }
 
     let size = slot.source_face.size_metrics();
-    let xstrength = ((FT_Long::from(size.x_ppem) * x_delta) / 1024) as i32;
-    let ystrength = ((FT_Long::from(size.y_ppem) * y_delta) / 1024) as i32;
+    let xstrength = (FT_Long::from(size.x_ppem) * x_delta) / 1024;
+    let ystrength = (FT_Long::from(size.y_ppem) * y_delta) / 1024;
     if slot.format == FT_GLYPH_FORMAT_OUTLINE {
-        slot.core_slot.adjust_outline_weight(xstrength, ystrength);
+        slot.core_slot
+            .adjust_outline_weight(xstrength as i32, ystrength as i32);
     } else {
         slot.core_slot.adjust_bitmap_weight(xstrength, ystrength);
     }
